Move Robot arm (zone3) to zone1
