+++ bsp/X300/multizone.cfg
@@ -9,6 +9,7 @@ Zone = 1 #
 	base = 0x10025000; size =    0x100; rwx = rw # PWM
 	base = 0x10012000; size =    0x100; rwx = rw # GPIO
 	base = 0x0C000000; size = 0x400000; rwx = rw # PLIC
+	base = 0x0200BFF8; size =   0x8; rwx = r  # RTC
 
 Zone = 2 #
 	base = 0x20420000; size =   64K; rwx = rx # FLASH
@@ -18,8 +19,7 @@ Zone = 2 #
 Zone = 3 #
 	base = 0x20430000; size =   64K; rwx = rx # FLASH
 	base = 0x80009000; size =    4K; rwx = rw # RAM
-	base = 0x0200BFF8; size =   0x8; rwx = r  # RTC
-	base = 0x10012000; size = 0x100; rwx = rw # GPIO
+
 
 Zone = 4 #
 	base = 0x20440000; size =   64K; rwx = rx # FLASH
